chore: rename `src` to `lib`
