+++ lib/trackingModules/palette/PaletteTracking.js
