+++ lib/trackingModules/palette/index.js
