+++ package.json
@@ -2,9 +2,9 @@
   "name": "bpmn-js-tracking",
   "version": "0.1.0",
   "description": "",
-  "main": "src/index.js",
+  "main": "lib/index.js",
   "files": [
-    "src"
+    "lib"
   ],
   "scripts": {
     "all": "run-s lint test",
